+++ src/main/java/treky/command/CommandHandler.java
@@ -4,8 +4,9 @@ import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 import java.util.List;
 
+import treky.storage.Storage;
 import treky.task.Task;
-import treky.TrekyException;
+import treky.exception.TrekyException;
 
 public class CommandHandler {
     private final TaskManager taskManager;
